LifeInsuranceDataModel.jl

The notion of section is used here to describe object revisions as of a certain pair of points in - 2 dimensional - transaction and reference time. So it is a geometrical metaphor of two crossing cuts = i.e. sections - reducing 2D-transaction and reference time to 1D-reference time using a point in transaction time and reducing 1D-reference time using a point in reference time to yield a point in the version space.

LifeInsuranceDataModel.PartnerSection — Type

PartnerSection

is a section (see above) of a Partner entity

LifeInsuranceDataModel.ProductItemSection — Type

ProductItemSection is a section (see above) of a ProductItem component

LifeInsuranceDataModel.Role — Type

Roleusin

role of a relationship

LifeInsuranceDataModel.TariffItemPartnerReference — Type

TariffItemPartnerReference is a reference from a TariffItem to a Partner entity For instance, typically an insured person

LifeInsuranceDataModel.TariffItemSection — Type

TariffItemSection is a section (see above) of a TariffItem component

LifeInsuranceDataModel.TariffItemTariffReference — Type

TariffItemTariffReference is a reference from a TariffItem to a Tariff entity

LifeInsuranceDataModel.TariffSection — Type

TariffSection

is a section (see above) of a Tariff entity